finished merging IncidentViewerPage -> LiveStreamPage
