--- frontend/src/features/pageContainer/pages/IncidentViewerPage.module.scss
@@ -1,32 +1,0 @@
-.pageContainer {
-}
-
-.incidentContainer {
-}
-
-.incidentContainerText {
-}
-
-.video {
-}
-
-.title {
-}
-
-.info {
-}
-
-.incidentRow {
-  display: flex;
-  justify-content: space-between;
-}
-
-.rowTitle {
-  font-weight: bold;
-}
-
-.rowData {
-}
-
-.incidentsBottom {
-}
