+++ frontend/src/features/pageContainer/pages/LiveStreamPage.module.scss
@@ -19,25 +19,38 @@
   padding: 32px;
 }
 
+.incidentRow {
+  display: flex;
+  justify-content: space-between;
+  padding-top: 15px;
+}
+
+.rowTitle {
+  font-weight: bold;
+}
+
+.button {
+  display: flex;
+  margin-top: auto;
+  width: 100%;
+  justify-content: center;
+}
+
 .containerLiveStream {
+  display: flex;
   padding: 32px;
   margin: 64px;
   margin-bottom: 16px;
   margin-top: 0px;
-  display: flex;
   flex-direction: row;
-  align-items: center;
+  justify-content: space-between;
 }
-
-// .containerVideo {
-//   padding-top: 48px;
-//   padding-left: 128px;
-// }
 
 .containerStreamData {
   padding: 16px;
   padding-top: 0;
   margin-right: 32px;
   height: 405px;
-  width: 784px;
+  width: 100%;
+  max-width: 400px;
 }
